Init module de tests
