+++ pom.xml
@@ -25,6 +25,7 @@
         <module>domain</module>
         <module>infrastructure</module>
         <module>bootstrap</module>
+        <module>tests/bdd</module>
     </modules>
 
     <properties>
@@ -70,6 +71,12 @@
             <dependency>
                 <groupId>org.hesperides</groupId>
                 <artifactId>infrastructure</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.hesperides</groupId>
+                <artifactId>bootstrap</artifactId>
                 <version>${project.version}</version>
             </dependency>
 
@@ -139,6 +146,12 @@
                 <groupId>org.hamcrest</groupId>
                 <artifactId>hamcrest-core</artifactId>
                 <version>${hamcrest-core.version}</version>
+                <scope>test</scope>
+            </dependency>
+            <dependency>
+                <groupId>org.springframework.security</groupId>
+                <artifactId>spring-security-test</artifactId>
+                <version>4.2.4.RELEASE</version>
                 <scope>test</scope>
             </dependency>
         </dependencies>
